backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Python backend with FastAPI"
authors = ["Michael Hua <michael.huayl@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
fastapi = {extras = ["standard"], version = "^0.115.0"}
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
alembic = "^1.13.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"
pytest = "^8.3.3"

[tool.ruff]
show-fixes = true
target-version = "py312"
line-length = 120

[tool.ruff.lint]
extend-select = ["N", "UP", "PL", "Q"]
ignore = ["E501"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
